v1.1: Items are now randomised within levels.
Custom level data with notes as to where items can be placed is now
included. See MAPINFO.TXT for info on how this works.

It's no-longer possible to exit a level without collecting its item.

A bug fix with the yorp messages, which were wrong about where levels
had moved.

The title screen's block colours have been randomised, and the main menu
has been simplified and updated with randomiser config info.

It's possible to run with the old randomiser rules with the /V1RULES
option, but it required the original LEVEL??.CK1 files, which are not
included.